feat. 게시글 목록, 상세 페이지 컴포넌트 분리 및 디자인
