+++ blog/src/components/Footer.tsx
@@ -1,0 +1,12 @@
+import {Link} from "react-router-dom";
+import React from "react";
+
+export default function Footer() {
+  return (
+    <footer>
+      <Link to="/posts/new">글쓰기</Link>
+      <Link to="/posts">게시글</Link>
+      <Link to="/profile">프로필</Link>
+    </footer>
+  )
+}
