+++ blog/src/components/Router.tsx
@@ -1,11 +1,11 @@
 import {Navigate, Route, Routes} from "react-router-dom";
 import React from "react";
 import Home from "../pages/home";
-import PostList from "../pages/posts";
+import PostsPage from "../pages/posts";
 import PostDetail from "../pages/posts/detail";
 import PostNew from "../pages/posts/new";
 import PostEdit from "../pages/posts/edit";
-import ProfilePage from "../pages/profile/profile";
+import ProfilePage from "../pages/profile";
 import LoginPage from "../pages/login";
 import SignupPage from "../pages/signup";
 
@@ -14,7 +14,7 @@ export default function Router() {
     <>
       <Routes>
         <Route path="/" element={<Home/>}/>
-        <Route path="/posts" element={<PostList/>}/>
+        <Route path="/posts" element={<PostsPage/>}/>
         <Route path="/posts/:id" element={<PostDetail/>}/>
         <Route path="/posts/new" element={<PostNew/>}/>
         <Route path="/posts/edit/:id" element={<PostEdit/>}/>
