+++ blog/src/pages/posts/detail.tsx
@@ -1,7 +1,14 @@
 import React from "react";
+import PostDetail from "../../components/PostDetail";
+import Header from "../../components/Header";
+import Footer from "../../components/Footer";
 
-export default function PostDetail() {
+export default function PostPage() {
   return (
-    <h1>Post Detail Page</h1>
+    <>
+      <Header />
+      <PostDetail/>
+      <Footer/>
+    </>
   );
 };
