+++ blog/src/pages/posts/index.tsx
@@ -1,7 +1,14 @@
 import React from "react";
+import Header from "../../components/Header";
+import Footer from "../../components/Footer";
+import PostList from "../../components/PostList";
 
-export default function PostList() {
+export default function PostsPage() {
   return (
-    <h1>Post List Page</h1>
+    <>
+      <Header/>
+      <PostList hasNavigation={false}/>
+      <Footer/>
+    </>
   );
 };
