+++ blog/src/pages/posts/new.tsx
@@ -1,7 +1,13 @@
 import React from "react";
+import Header from "../../components/Header";
+import Footer from "../../components/Footer";
 
 export default function PostNew() {
   return (
-    <h1>Post New Page</h1>
+    <>
+      <Header/>
+      <div>Post Form</div>
+      <Footer/>
+    </>
   );
 };
