+++ blog/src/pages/profile/index.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import Profile from "../../components/Profile";
+import Header from "../../components/Header";
+import Footer from "../../components/Footer";
+import PostList from "../../components/PostList";
+
+export default function ProfilePage() {
+  return <>
+    <Header/>
+    <Profile/>
+    <PostList hasNavigation={false}/>
+    <Footer/>
+  </>
+};
