--- blog/src/pages/profile/profile.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-export default function ProfilePage() {
-  return (
-    <h1>Profile Page</h1>
-  );
-};
